requirements.txt
experta==1.9.4
streamlit==1.28.0
pyyaml==6.0.1
pandas>=2.1.0
numpy>=1.24.3
